+++ uCOS-III_Wrapper/uCOS-III/os.h
@@ -345,7 +345,7 @@ typedef  enum  os_err {
     OS_ERR_NONE                      =     0u,
 
     OS_ERR_A                         = 10000u,
-//    OS_ERR_ACCEPT_ISR                = 10001u,
+    OS_ERR_ACCEPT_ISR                = 10001u, /*原版3.03中仅定义未使用*/
 
     OS_ERR_B                         = 11000u,
 
@@ -360,9 +360,9 @@ typedef  enum  os_err {
     OS_ERR_F                         = 15000u,
     OS_ERR_FATAL_RETURN              = 15001u,
 
-//    OS_ERR_FLAG_GRP_DEPLETED         = 15101u,
-//    OS_ERR_FLAG_NOT_RDY              = 15102u,
-//    OS_ERR_FLAG_PEND_OPT             = 15103u,
+    OS_ERR_FLAG_GRP_DEPLETED         = 15101u, /*原版3.03中仅定义未使用*/
+    OS_ERR_FLAG_NOT_RDY              = 15102u, /*原版3.03中仅定义未使用*/
+    OS_ERR_FLAG_PEND_OPT             = 15103u,
     OS_ERR_FLUSH_ISR                 = 15104u,
 
     OS_ERR_G                         = 16000u,
@@ -382,7 +382,7 @@ typedef  enum  os_err {
     OS_ERR_K                         = 20000u,
 
     OS_ERR_L                         = 21000u,
-//    OS_ERR_LOCK_NESTING_OVF          = 21001u,
+    OS_ERR_LOCK_NESTING_OVF          = 21001u,
 
     OS_ERR_M                         = 22000u,
 
@@ -401,8 +401,8 @@ typedef  enum  os_err {
 //    OS_ERR_MSG_POOL_NULL_PTR         = 22302u,
 
     OS_ERR_MUTEX_NOT_OWNER           = 22401u,
-//    OS_ERR_MUTEX_OWNER               = 22402u,
-//    OS_ERR_MUTEX_NESTING             = 22403u,
+    OS_ERR_MUTEX_OWNER               = 22402u,
+    OS_ERR_MUTEX_NESTING             = 22403u,
 
     OS_ERR_N                         = 23000u,
     OS_ERR_NAME                      = 23001u,
@@ -424,22 +424,22 @@ typedef  enum  os_err {
     OS_ERR_PEND_ABORT_ISR            = 25002u,
     OS_ERR_PEND_ABORT_NONE           = 25003u,
     OS_ERR_PEND_ABORT_SELF           = 25004u,
-//    OS_ERR_PEND_DEL                  = 25005u,
+    OS_ERR_PEND_DEL                  = 25005u, /*原版3.03中仅定义未使用*/
     OS_ERR_PEND_ISR                  = 25006u,
-//    OS_ERR_PEND_LOCKED               = 25007u,
+    OS_ERR_PEND_LOCKED               = 25007u, /*原版3.03中仅定义未使用*/
 //    OS_ERR_PEND_WOULD_BLOCK          = 25008u,
 
-//    OS_ERR_POST_NULL_PTR             = 25101u,
+    OS_ERR_POST_NULL_PTR             = 25101u, /*原版3.03中仅定义未使用*/
     OS_ERR_POST_ISR                  = 25102u,
 
-//    OS_ERR_PRIO_EXIST                = 25201u,
-//    OS_ERR_PRIO                      = 25202u,
+    OS_ERR_PRIO_EXIST                = 25201u, /*原版3.03中仅定义未使用*/
+    OS_ERR_PRIO                      = 25202u, /*原版3.03中仅定义未使用*/
     OS_ERR_PRIO_INVALID              = 25203u,
 
     OS_ERR_PTR_INVALID               = 25301u,
 
     OS_ERR_Q                         = 26000u,
-//    OS_ERR_Q_FULL                    = 26001u,
+    OS_ERR_Q_FULL                    = 26001u, /*原版3.03中仅定义未使用*/
 //    OS_ERR_Q_EMPTY                   = 26002u,
     OS_ERR_Q_MAX                     = 26003u,
     OS_ERR_Q_SIZE                    = 26004u,
@@ -447,16 +447,16 @@ typedef  enum  os_err {
     OS_ERR_R                         = 27000u,
     OS_ERR_REG_ID_INVALID            = 27001u,
 //    OS_ERR_ROUND_ROBIN_1             = 27002u,
-//    OS_ERR_ROUND_ROBIN_DISABLED      = 27003u,
+    OS_ERR_ROUND_ROBIN_DISABLED      = 27003u,
 
     OS_ERR_S                         = 28000u,
-//    OS_ERR_SCHED_INVALID_TIME_SLICE  = 28001u,
+    OS_ERR_SCHED_INVALID_TIME_SLICE  = 28001u, /*原版3.03中仅定义未使用*/
     OS_ERR_SCHED_LOCK_ISR            = 28002u,
     OS_ERR_SCHED_LOCKED              = 28003u,
     OS_ERR_SCHED_NOT_LOCKED          = 28004u,
     OS_ERR_SCHED_UNLOCK_ISR          = 28005u,
 
-//    OS_ERR_SEM_OVF                   = 28101u,
+    OS_ERR_SEM_OVF                   = 28101u,
     OS_ERR_SET_ISR                   = 28102u,
 
     OS_ERR_STAT_RESET_ISR            = 28201u,
@@ -467,31 +467,31 @@ typedef  enum  os_err {
 //    OS_ERR_STATUS_INVALID            = 28206u,
     OS_ERR_STK_INVALID               = 28207u,
     OS_ERR_STK_SIZE_INVALID          = 28208u,
-//    OS_ERR_STK_LIMIT_INVALID         = 28209u,
+    OS_ERR_STK_LIMIT_INVALID         = 28209u,
 
     OS_ERR_T                         = 29000u,
 //    OS_ERR_TASK_CHANGE_PRIO_ISR      = 29001u,
     OS_ERR_TASK_CREATE_ISR           = 29002u,
-//    OS_ERR_TASK_DEL                  = 29003u,
+    OS_ERR_TASK_DEL                  = 29003u, /*原版3.03中仅定义未使用*/
 //    OS_ERR_TASK_DEL_IDLE             = 29004u,
 //    OS_ERR_TASK_DEL_INVALID          = 29005u,
     OS_ERR_TASK_DEL_ISR              = 29006u,
     OS_ERR_TASK_INVALID              = 29007u,
-//    OS_ERR_TASK_NO_MORE_TCB          = 29008u,
+    OS_ERR_TASK_NO_MORE_TCB          = 29008u, /*原版3.03中仅定义未使用*/
     OS_ERR_TASK_NOT_DLY              = 29009u,
     OS_ERR_TASK_NOT_EXIST            = 29010u,
     OS_ERR_TASK_NOT_SUSPENDED        = 29011u,
     OS_ERR_TASK_OPT                  = 29012u,
     OS_ERR_TASK_RESUME_ISR           = 29013u,
-//    OS_ERR_TASK_RESUME_PRIO          = 29014u,
+    OS_ERR_TASK_RESUME_PRIO          = 29014u, /*原版3.03中仅定义未使用*/
     OS_ERR_TASK_RESUME_SELF          = 29015u,
-//    OS_ERR_TASK_RUNNING              = 29016u,
+    OS_ERR_TASK_RUNNING              = 29016u, /*原版3.03中仅定义未使用*/
     OS_ERR_TASK_STK_CHK_ISR          = 29017u,
     OS_ERR_TASK_SUSPENDED            = 29018u,
 //    OS_ERR_TASK_SUSPEND_IDLE         = 29019u,
 //    OS_ERR_TASK_SUSPEND_INT_HANDLER  = 29020u,
     OS_ERR_TASK_SUSPEND_ISR          = 29021u,
-//    OS_ERR_TASK_SUSPEND_PRIO         = 29022u,
+    OS_ERR_TASK_SUSPEND_PRIO         = 29022u, /*原版3.03中仅定义未使用*/
     OS_ERR_TASK_WAITING              = 29023u,
 
     OS_ERR_TCB_INVALID               = 29101u,
@@ -509,26 +509,26 @@ typedef  enum  os_err {
 
     OS_ERR_TIME_DLY_ISR              = 29301u,
     OS_ERR_TIME_DLY_RESUME_ISR       = 29302u,
-//    OS_ERR_TIME_GET_ISR              = 29303u,
+    OS_ERR_TIME_GET_ISR              = 29303u, /*原版3.03中仅定义未使用*/
     OS_ERR_TIME_INVALID_HOURS        = 29304u,
     OS_ERR_TIME_INVALID_MINUTES      = 29305u,
     OS_ERR_TIME_INVALID_SECONDS      = 29306u,
     OS_ERR_TIME_INVALID_MILLISECONDS = 29307u,
-//    OS_ERR_TIME_NOT_DLY              = 29308u,
-//    OS_ERR_TIME_SET_ISR              = 29309u,
+    OS_ERR_TIME_NOT_DLY              = 29308u, /*原版3.03中仅定义未使用*/
+    OS_ERR_TIME_SET_ISR              = 29309u, /*原版3.03中仅定义未使用*/
     OS_ERR_TIME_ZERO_DLY             = 29310u,
 
     OS_ERR_TIMEOUT                   = 29401u,
 
 //    OS_ERR_TMR_INACTIVE              = 29501u,
-//    OS_ERR_TMR_INVALID_DEST          = 29502u,
+    OS_ERR_TMR_INVALID_DEST          = 29502u, /*原版3.03中仅定义未使用*/
     OS_ERR_TMR_INVALID_DLY           = 29503u,
     OS_ERR_TMR_INVALID_PERIOD        = 29504u,
     OS_ERR_TMR_INVALID_STATE         = 29505u,
     OS_ERR_TMR_INVALID               = 29506u,
     OS_ERR_TMR_ISR                   = 29507u,
     OS_ERR_TMR_NO_CALLBACK           = 29508u,
-//    OS_ERR_TMR_NON_AVAIL             = 29509u,
+    OS_ERR_TMR_NON_AVAIL             = 29509u, /*原版3.03中仅定义未使用*/
 //    OS_ERR_TMR_PRIO_INVALID          = 29510u,
 //    OS_ERR_TMR_STK_INVALID           = 29511u,
 //    OS_ERR_TMR_STK_SIZE_INVALID      = 29512u,